Moved most of dispatcher-agent
This is an attempt to speed up compilation and testing time.
