+++ rust/controller-agent/src/compiler.rs
@@ -52,7 +52,7 @@ fn gen_function_table_file(table: &[String]) -> TokenStream {
     let q_inserts: Vec<_> = table.iter().map(|x| gen_function_table_entry(x)).collect();
     return quote! {
         // This file is generated by the compiler in controller-agent.
-        use super::trace_runtime::Containerless;
+        use dispatcher_agent_lib::trace_runtime::Containerless;
         use std::collections::HashMap;
         #(#q_mods)*
