+++ rust/dispatcher-agent-lib/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "dispatcher-agent-lib"
+version = "0.1.0"
+authors = ["Arjun Guha <arjun@cs.umass.edu>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+shared2 = { path = "../shared2" }
+tokio = { version = "*", features = ["macros", "sync", "signal", "time"] }
+hyper = "*"
+kube = { version = "0.25.0", features = ["openapi"] }
+k8s-openapi = { version = "0.7.1", default-features = false, features = ["v1_17"] }
+serde_json = "*"
+futures = "*"
+http = "*"
+bytes = "*"
+futures-retry = "*"
+k8s = { path = "../k8s" }
+lazy_static = "*"
+regex = "*"
+log = "*"
+bumpalo = { version = "*", features = ["collections"] }
+thiserror = "*"
+warp = "*"
