+++ rust/dispatcher-agent-lib/src/autoscaler.rs
