+++ rust/dispatcher-agent-lib/src/error.rs
