+++ rust/dispatcher-agent-lib/src/function_manager.rs
