+++ rust/dispatcher-agent-lib/src/state.rs
