+++ rust/dispatcher-agent-lib/src/types.rs
