+++ rust/dispatcher-agent-lib/src/util.rs
