+++ rust/dispatcher-agent-lib/src/windowed_max.rs
