+++ rust/dispatcher-agent/src/decontainerized_functions/mod.rs
@@ -1,6 +1,6 @@
-use super::trace_runtime::Containerless;
+use dispatcher_agent_lib::trace_runtime::Containerless;
 use std::collections::HashMap;
 pub fn init() -> HashMap<&'static str, Containerless> {
-    let mut ht: HashMap<&'static str, Containerless> = HashMap::new();
+    let ht: HashMap<&'static str, Containerless> = HashMap::new();
     return ht;
 }
